Added interface and service class
