+++ src/main/java/com/game/design/cardgame/service/DeckService.java
@@ -1,0 +1,8 @@
+package com.game.design.cardgame.service;
+
+import com.game.design.cardgame.entity.Card;
+
+public interface DeckService {
+    void shuffleDeck();
+    Card drawCard();
+}
